feat : 일정 미리보기
